+++ helperfuncs.c
@@ -2,7 +2,8 @@
  * Helper functions for rcopy and server packet creation
  * File: helperfuncs.c
  * 
- * Program Description: This file contains helper functions for creating PDUs
+ * Program Description: 
+ * This file contains helper functions for creating PDUs for the use of rcopy and server.
  * Author: Elizabeth Acevedo
  * Date created: 03/11/2025
  * 
@@ -12,7 +13,7 @@
 #include "helperfuncs.h"
 #include "cpe464.h"
 
-void createPDU(uint8_t *pdu, uint8_t *data_chunk, uint16_t data_size, uint32_t packet_num, uint8_t flag) {
+int createPDU(uint8_t *pdu, uint8_t *data_chunk, uint16_t data_size, uint32_t packet_num, uint8_t flag) {
 	// Convert packet number to network order
 	uint32_t packet_num_netword = htonl(packet_num);
 
@@ -33,5 +34,21 @@ void createPDU(uint8_t *pdu, uint8_t *data_chunk, uint16_t data_size, uint32_t p
     checksum = in_cksum((unsigned short *)pdu, data_size + 7);
     memcpy(&pdu[4], &checksum, 2);
 
+	return data_size + 7; // Return the size of the PDU (data size + header size)
 }
 
+
+float getErrorRate(char *errate_entry) {
+	// Convert error rate entry to float
+	char *endptr;
+	float errorRate = strtof(errate_entry, &endptr);
+
+	// Check for conversion errors
+	if (errate_entry == endptr) {
+        	printf("Error: '%s' is not a valid float.\n", errate_entry);
+        	return -1;
+    }
+
+    return errorRate;
+
+}