+++ networks.c
@@ -19,6 +19,7 @@
 #include "networks.h"
 #include "gethostbyname.h"
 
+struct connectionInfo *connect_info;
 
 // This funciton creates a UDP socket on the server side and binds to that socket.  
 // It prints out the port number and returns the socket number.
@@ -91,4 +92,39 @@ int setupUdpClientToServer(struct sockaddr_in6 *serverAddress, char * hostName, 
 	return socketNum;
 }
 
+void setupConnectionInfo(uint8_t *from_filename, uint8_t *to_filename, uint32_t window_size, uint16_t buffer_size, float errorRate, char *hostName, int serverPort, struct sockaddr_in6 *serverAddress) {
+	// Allocate memory for connectionInfo struct
+	connect_info = (struct connectionInfo *)malloc(sizeof(struct connectionInfo));
+	if (connect_info == NULL) {
+		perror("Failed to allocate memory for connectionInfo");
+		exit(EXIT_FAILURE);
+	}
 
+	// Set the fields of the connectionInfo struct
+	connect_info->from_filename = from_filename;
+	connect_info->to_filename = to_filename;
+	connect_info->window_size = window_size;
+	connect_info->buffer_size = buffer_size;
+	connect_info->errorRate = errorRate;
+	connect_info->hostName = hostName;
+	connect_info->serverPort = serverPort;
+	connect_info->serverAddress = serverAddress;
+}
+
+struct connectionInfo getConnectionInfo() {
+	return *connect_info;
+}
+
+int replaceSocket(int old_socket) {
+	int new_socket = 0;
+	close(old_socket);
+	new_socket = setupUdpClientToServer(connect_info->serverAddress, connect_info->hostName, connect_info->serverPort);
+	return new_socket;
+}
+
+void freeConnectInfo() {
+	// Free the memory allocated for connectionInfo struct
+	if (connect_info) {
+		free(connect_info);
+	}
+}