+++ rcopyCBuffer.h
@@ -32,10 +32,7 @@ struct packetinBuffer {
 void initBuffer(int window_size);
 int getPacketNum(uint8_t *packet);
 void storetoBuffer(uint8_t *packet, uint32_t packet_size);
-void writePDUtoDisk(uint32_t packet_num, uint8_t *leaving_packet, uint32_t buff_indx);
-void retrieveNextPacket(uint8_t *packet);
-void testStorePkt(int seq_num, int size);
-void printBuffer();
+int retrieveNextPacket(uint8_t *packet);
 void freeRcopyBuffer();